major refactor to remove `KSP`
