+++ gradle.properties
@@ -6,10 +6,8 @@ kotlin.compiler.execution.strategy=in-process
 
 GROUP_ID=dev.arildo
 LIB_ARTIFACT_ID=iris-mock
-LIB_VERSION=1.0.10
+LIB_VERSION=1.0.1
 
 PLUGIN_ID=dev.arildo.iris-mock-plugin
-PLUGIN_VERSION=1.0.10
-
-COMPILER_ARTIFACT_ID=iris-mock-compiler
-COMPILER_VERSION=1.0.10
+PLUGIN_ARTIFACT_ID=iris-mock-plugin
+PLUGIN_VERSION=1.0.1
