+++ iris-mock-plugin/build.gradle.kts
@@ -2,6 +2,7 @@ import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
 plugins {
     id("com.gradle.plugin-publish") version "1.2.0"
+    id("com.github.gmazzo.buildconfig") version "3.0.3"
     `kotlin-dsl`
     signing
     kotlin("jvm")
@@ -47,8 +48,8 @@ dependencies {
     compileOnly("dev.gradleplugins:gradle-api:7.6")
     compileOnly("com.squareup.okhttp3:okhttp:3.14.9")
     compileOnly("org.jetbrains.kotlin:kotlin-compiler-embeddable:1.8.21")
-    compileOnly("com.google.auto.service:auto-service-annotations:1.0")
-    kapt("com.google.auto.service:auto-service:1.0")
+    compileOnly("com.google.auto.service:auto-service-annotations:1.0.1")
+    kapt("com.google.auto.service:auto-service:1.0.1")
 
     arrayOf("asm", "asm-util", "asm-commons").forEach {
         compileOnly("org.ow2.asm:$it:9.4")
@@ -59,12 +60,6 @@ tasks.withType<KotlinCompile> {
     kotlinOptions.jvmTarget = "1.8"
 }
 
-kotlin {
-    jvmToolchain {
-        languageVersion.set(JavaLanguageVersion.of(11))
-    }
-}
-
 tasks.withType<KotlinCompile> {
     kotlinOptions {
         freeCompilerArgs += listOf(
@@ -72,4 +67,15 @@ tasks.withType<KotlinCompile> {
             "-opt-in=org.jetbrains.kotlin.compiler.plugin.ExperimentalCompilerApi"
         )
     }
+}
+
+val pluginId = findProperty("PLUGIN_ID").toString()
+val pluginArtifactId = findProperty("PLUGIN_ARTIFACT_ID").toString()
+
+buildConfig {
+    packageName("dev.arildo.iris.plugin")
+    buildConfigField("String", "PLUGIN_ID", "\"${pluginId}\"")
+    buildConfigField("String", "PLUGIN_GROUP_ID", "\"${group}\"")
+    buildConfigField("String", "PLUGIN_ARTIFACT_ID", "\"${pluginArtifactId}\"")
+    buildConfigField("String", "PLUGIN_VERSION", "\"${version}\"")
 }
