+++ iris-mock-plugin/src/main/kotlin/dev/arildo/iris/plugin/WrapperCodeGenerator.kt
@@ -1,22 +1,23 @@
 package dev.arildo.iris.plugin
 
 import com.google.auto.service.AutoService
-import dev.arildo.iris.plugin.util.ClassReference
-import dev.arildo.iris.plugin.util.CodeGenerator
+import dev.arildo.iris.plugin.codegen.ClassReference
+import dev.arildo.iris.plugin.codegen.CodeGenerator
+import dev.arildo.iris.plugin.codegen.classAndInnerClassReferences
+import dev.arildo.iris.plugin.codegen.generateIrisMockWrapper
+import dev.arildo.iris.plugin.codegen.irisMockWrapperFactory
 import dev.arildo.iris.plugin.util.IRIS_MOCK_INTERCEPTOR
 import dev.arildo.iris.plugin.util.IRIS_WRAPPER_NAME
 import dev.arildo.iris.plugin.util.IRIS_WRAPPER_PACKAGE
-import dev.arildo.iris.plugin.util.classAndInnerClassReferences
-import dev.arildo.iris.plugin.util.createGeneratedFile
 import dev.arildo.iris.plugin.util.fq
 import org.jetbrains.kotlin.descriptors.ModuleDescriptor
 import org.jetbrains.kotlin.psi.KtFile
 import java.io.File
 
 @AutoService(CodeGenerator::class)
-internal class WrapperCodeGenerator : PrivateCodeGenerator() {
+internal class WrapperCodeGenerator : CodeGenerator {
 
-    override fun generateCodePrivate(
+    override fun generateCode(
         codeGenDir: File,
         module: ModuleDescriptor,
         projectFiles: Collection<KtFile>
@@ -25,19 +26,15 @@ internal class WrapperCodeGenerator : PrivateCodeGenerator() {
             .classAndInnerClassReferences(module)
             .filter { it.isAnnotatedWithIrisMock() }
 
-        createGeneratedFile(
+        generateIrisMockWrapper(
             codeGenDir = codeGenDir,
             packageName = IRIS_WRAPPER_PACKAGE,
             fileName = IRIS_WRAPPER_NAME,
-            content = wrapperInterceptorFactory(annotatedClasses)
+            content = irisMockWrapperFactory(annotatedClasses)
         )
     }
-
 }
 
-private fun ClassReference.Psi.isAnnotatedWithIrisMock(): Boolean {
-    return annotations.any {
-        it.annotation.fq(module)?.fqName?.asString()?.contains(IRIS_MOCK_INTERCEPTOR)
-            ?: false
-    }
+private fun ClassReference.Psi.isAnnotatedWithIrisMock() = annotations.any {
+    it.annotation.fq(module)?.fqName?.asString()?.contains(IRIS_MOCK_INTERCEPTOR) ?: false
 }
