+++ iris-mock-plugin/src/main/kotlin/dev/arildo/iris/plugin/codegen/AnnotatedReference.kt
@@ -1,0 +1,8 @@
+package dev.arildo.iris.plugin.codegen
+
+/**
+ * Adapted from https://github.com/square/anvil
+ */
+interface AnnotatedReference {
+  val annotations: List<AnnotationReference>
+}
