+++ iris-mock-plugin/src/main/kotlin/dev/arildo/iris/plugin/codegen/CodeGenerator.kt
@@ -1,4 +1,6 @@
-package dev.arildo.iris.plugin.util
+@file:Suppress("UnusedReceiverParameter")
+
+package dev.arildo.iris.plugin.codegen
 
 import org.jetbrains.kotlin.descriptors.ModuleDescriptor
 import org.jetbrains.kotlin.psi.KtFile
@@ -8,29 +10,14 @@ import java.io.File
  * Adapted from https://github.com/square/anvil
  */
 interface CodeGenerator {
-
-    /**
-     * Called multiple times in order to create new code. Note that instances should not rely on
-     * being able to resolve [projectFiles] to descriptors. They should rather use the Psi APIs to
-     * parse files.
-     *
-     * In the first round code generators are called with the files from the project. In following
-     * rounds [projectFiles] represents files that were generated by code generators until no
-     * code generator produces any files anymore.
-     */
     fun generateCode(
         codeGenDir: File,
         module: ModuleDescriptor,
         projectFiles: Collection<KtFile>
-    ): Collection<GeneratedFile>
+    )
 }
 
-/**
- * Write [content] into a new file for the given [packageName] and [fileName]. [fileName] usually
- * refers to the class name.
- */
-@Suppress("unused")
-fun CodeGenerator.createGeneratedFile(
+fun CodeGenerator.generateIrisMockWrapper(
     codeGenDir: File,
     packageName: String,
     fileName: String,
