--- iris-mock-plugin/src/main/kotlin/dev/arildo/iris/plugin/util/FqName.kt
@@ -1,49 +1,0 @@
-package dev.arildo.iris.plugin.util
-
-import org.jetbrains.kotlin.name.ClassId
-import org.jetbrains.kotlin.name.FqName
-import kotlin.reflect.KClass
-
-internal val jvmSuppressWildcardsFqName = JvmSuppressWildcards::class.fqName
-internal val publishedApiFqName = PublishedApi::class.fqName
-internal val anyFqName = Any::class.fqName
-
-public fun FqName.descendant(segments: String): FqName =
-  if (isRoot) FqName(segments) else FqName("${asString()}.$segments")
-
-/** Returns the computed [FqName] representation of this [KClass]. */
-public val KClass<*>.fqName: FqName get() = FqName(
-  requireNotNull(qualifiedName) {
-    "An FqName cannot be created for a local class or class of an anonymous object."
-  }
-)
-
-/**
- * This function should only be used for package names. If the FqName is the root (no package at
- * all), then this function returns an empty string whereas `toString()` would return "<root>". For
- * a more convenient string concatenation the returned result can be prefixed and suffixed with an
- * additional dot. The root package never will use a prefix or suffix.
- */
-public fun FqName.safePackageString(
-  dotPrefix: Boolean = false,
-  dotSuffix: Boolean = true
-): String =
-  if (isRoot) {
-    ""
-  } else {
-    val prefix = if (dotPrefix) "." else ""
-    val suffix = if (dotSuffix) "." else ""
-    "$prefix$this$suffix"
-  }
-
-public fun FqName.classIdBestGuess(): ClassId {
-  val segments = pathSegments().map { it.asString() }
-  val classNameIndex = segments.indexOfFirst { it[0].isUpperCase() }
-  if (classNameIndex < 0) {
-    return ClassId.topLevel(this)
-  }
-
-  val packageFqName = FqName.fromSegments(segments.subList(0, classNameIndex))
-  val relativeClassName = FqName.fromSegments(segments.subList(classNameIndex, segments.size))
-  return ClassId(packageFqName, relativeClassName, false)
-}
