+++ iris-mock-plugin/src/main/kotlin/dev/arildo/iris/plugin/util/FqNameExtensions.kt
@@ -1,0 +1,23 @@
+package dev.arildo.iris.plugin.util
+
+import org.jetbrains.kotlin.name.ClassId
+import org.jetbrains.kotlin.name.FqName
+
+/**
+ * Adapted from https://github.com/square/anvil
+ */
+fun FqName.descendant(segments: String): FqName =
+  if (isRoot) FqName(segments) else FqName("${asString()}.$segments")
+
+
+fun FqName.classIdBestGuess(): ClassId {
+  val segments = pathSegments().map { it.asString() }
+  val classNameIndex = segments.indexOfFirst { it[0].isUpperCase() }
+  if (classNameIndex < 0) {
+    return ClassId.topLevel(this)
+  }
+
+  val packageFqName = FqName.fromSegments(segments.subList(0, classNameIndex))
+  val relativeClassName = FqName.fromSegments(segments.subList(classNameIndex, segments.size))
+  return ClassId(packageFqName, relativeClassName, false)
+}
