--- iris-mock-plugin/src/main/kotlin/dev/arildo/iris/plugin/util/KotlinPoetUtils.kt
@@ -1,135 +1,0 @@
-@file:Suppress("unused")
-
-package dev.arildo.iris.plugin.util
-
-import com.squareup.kotlinpoet.AnnotationSpec
-import com.squareup.kotlinpoet.ClassName
-import com.squareup.kotlinpoet.FileSpec
-import com.squareup.kotlinpoet.TypeName
-import com.squareup.kotlinpoet.jvm.jvmSuppressWildcards
-import org.jetbrains.kotlin.descriptors.ClassDescriptor
-import org.jetbrains.kotlin.descriptors.ModuleDescriptor
-import org.jetbrains.kotlin.descriptors.PackageFragmentDescriptor
-import org.jetbrains.kotlin.name.ClassId
-import org.jetbrains.kotlin.name.FqName
-import org.jetbrains.kotlin.resolve.descriptorUtil.parents
-import org.jetbrains.kotlin.resolve.descriptorUtil.parentsWithSelf
-import java.io.ByteArrayOutputStream
-
-public fun ClassDescriptor.asClassName(): ClassName =
-  ClassName(
-    packageName = parents.filterIsInstance<PackageFragmentDescriptor>().first()
-      .fqName.safePackageString(dotSuffix = false),
-    simpleNames = parentsWithSelf.filterIsInstance<ClassDescriptor>()
-      .map { it.name.asString() }
-      .toList()
-      .reversed()
-  )
-
-public fun FqName.asClassName(module: ModuleDescriptor): ClassName {
-  return asClassNameOrNull(module)
-    ?: throw Exception("Couldn't parse ClassName for $this.")
-}
-
-private fun FqName.asClassNameOrNull(module: ModuleDescriptor): ClassName? {
-  val segments = pathSegments().map { it.asString() }
-
-  // If the first sentence case is not the last segment of the path it becomes ambiguous,
-  // for example, com.Foo.Bar could be an inner class Bar or an unconventional package com.Foo.
-  val canGuessClassName = segments.indexOfFirst { it[0].isUpperCase() } == segments.size - 1
-  if (canGuessClassName) {
-    return ClassName.bestGuess(asString())
-  }
-
-  for (index in (segments.size - 1) downTo 1) {
-    val packageSegments = segments.subList(0, index)
-    val classSegments = segments.subList(index, segments.size)
-
-    val validFqName = FqName.fromSegments(packageSegments + classSegments)
-      .canResolveFqName(module)
-
-    if (validFqName) {
-      return ClassName(
-        packageName = packageSegments.joinToString(separator = "."),
-        simpleNames = classSegments
-      )
-    }
-  }
-
-  // No matching class found. It's either a package name only or doesn't exist.
-  return null
-}
-
-public fun ClassId.asClassName(): ClassName {
-  return ClassName(
-    packageName = packageFqName.asString(),
-    simpleNames = relativeClassName.pathSegments().map { it.asString() }
-  )
-}
-
-public fun TypeName.withJvmSuppressWildcardsIfNeeded(
-  annotatedReference: AnnotatedReference,
-  typeReference: TypeReference
-): TypeName {
-  // If the parameter is annotated with @JvmSuppressWildcards, then add the annotation
-  // to our type so that this information is forwarded when our Factory is compiled.
-  val hasJvmSuppressWildcards = annotatedReference.isAnnotatedWith(jvmSuppressWildcardsFqName)
-
-  // Add the @JvmSuppressWildcards annotation even for simple generic return types like
-  // Set<String>. This avoids some edge cases where Dagger chokes.
-  val isGenericType = typeReference.isGenericType()
-
-  // Same for functions.
-  val isFunctionType = typeReference.isFunctionType()
-
-  return when {
-    hasJvmSuppressWildcards || isGenericType -> this.jvmSuppressWildcards()
-    isFunctionType -> this.jvmSuppressWildcards()
-    else -> this
-  }
-}
-
-private fun FileSpec.writeToString(): String {
-  val stream = ByteArrayOutputStream()
-  stream.writer().use {
-    writeTo(it)
-  }
-  return stream.toString()
-}
-
-private fun FileSpec.Builder.suppressWarnings() {
-  addAnnotation(
-    AnnotationSpec
-      .builder(Suppress::class)
-      // Suppress deprecation warnings.
-      .addMember("\"DEPRECATION\"")
-      // Suppress errors for experimental features in generated code.
-      .apply {
-        if (KotlinVersion.CURRENT > KotlinVersion(1, 6, 10)) {
-          addMember("\"OPT_IN_USAGE\"")
-          addMember("\"OPT_IN_USAGE_ERROR\"")
-        } else {
-          addMember("\"EXPERIMENTAL_API_USAGE_ERROR\"")
-          addMember("\"EXPERIMENTAL_API_USAGE\"")
-        }
-      }
-      .build()
-  )
-}
-
-public fun FileSpec.Companion.buildFile(
-  packageName: String,
-  fileName: String,
-  generatorComment: String = "Generated by Anvil.\nhttps://github.com/square/anvil",
-  block: FileSpec.Builder.() -> Unit
-): String =
-  builder(packageName, fileName)
-    .apply {
-      // Suppress any deprecation warnings.
-      suppressWarnings()
-
-      block()
-    }
-    .addFileComment(generatorComment)
-    .build()
-    .writeToString()
