--- iris-mock-plugin/src/main/kotlin/dev/arildo/iris/plugin/util/Visibility.kt
@@ -1,8 +1,0 @@
-package dev.arildo.iris.plugin.util
-
-public enum class Visibility {
-    PUBLIC,
-    INTERNAL,
-    PROTECTED,
-    PRIVATE
-}
