--- iris-mock-plugin/src/main/resources/META-INF/services/org.jetbrains.kotlin.compiler.plugin.CompilerPluginRegistrar
@@ -1,1 +1,0 @@
-dev.arildo.iris.plugin.IrisMockComponentRegistrar
