+++ iris-mock/src/main/kotlin/dev/arildo/iris/mock/IrisMock.kt
@@ -3,5 +3,4 @@ package dev.arildo.iris.mock
 object IrisMock {
     var enableLogs = true
         internal set
-
 }
