+++ iris-mock/src/main/kotlin/dev/arildo/iris/mock/annotation/IrisMockInterceptor.kt
@@ -1,3 +1,5 @@
+@file:Suppress("unused")
+
 package dev.arildo.iris.mock.annotation
 
 @Target(AnnotationTarget.CLASS)
