+++ sample/proguard-rules.pro
@@ -9,7 +9,7 @@
 # and specify the fully qualified class name to the JavaScript interface
 # class:
 #-keepclassmembers class fqcn.of.javascript.interface.for.webview {
-#   public *;
+#   *;
 #}
 
 # Uncomment this to preserve the line number information for
